Add support for missing za4 features
This is a bit hacky. The services from miio_fan likely need changing depending on how the fan is set up (and for multiple fans etc). However, for default configurations it should work okay.

 - Adds support to set oscillation angle
 - Adds support to set child lock on/off
 - Updates cursor to be a pointer when rotating the fan

Could also add a new icon to turn buzzer on and off, but that would require adding a new element to the UI.

The big thing that is missing is setting the timer. This appears to be a limitation with the miio library so I haven't tackled that yet. (Of course with HA you can create your own timer).