Added support for dmaker.fan.p5 model
p5 model does not report current direct_speed at which it's operating, so need to use Levels 1-4 which miio reports back to us.
Because p5 reporting only 4 levels, need to loop from 4th level back to 1st. That's the reason levels are different from za4 as well.
In order to support p5 and za4 simultaniusly, added speed_list array.
According to my plan, it should work for za4 as before.